Fix custom inbound PUT and GET paths
