+++ components/org.wso2.carbon.identity.api.server.application.management/org.wso2.carbon.identity.api.server.application.management.v1/src/main/java/org/wso2/carbon/identity/api/server/application/management/v1/core/functions/application/inbound/WSTrustInboundUtils.java
@@ -15,12 +15,17 @@
  */
 package org.wso2.carbon.identity.api.server.application.management.v1.core.functions.application.inbound;
 
+import org.apache.commons.lang.StringUtils;
 import org.wso2.carbon.identity.api.server.application.management.common.ApplicationManagementServiceHolder;
 import org.wso2.carbon.identity.api.server.application.management.v1.WSTrustConfiguration;
 import org.wso2.carbon.identity.api.server.application.management.v1.core.functions.Utils;
 import org.wso2.carbon.identity.application.authentication.framework.util.FrameworkConstants.StandardInboundProtocols;
+import org.wso2.carbon.identity.application.common.model.InboundAuthenticationRequestConfig;
 import org.wso2.carbon.identity.application.common.model.ServiceProvider;
 import org.wso2.carbon.security.SecurityConfigException;
+import org.wso2.carbon.security.sts.service.util.TrustedServiceData;
+
+import java.util.Arrays;
 
 import static org.wso2.carbon.identity.api.server.application.management.v1.core.functions.Utils.buildServerErrorResponse;
 
@@ -54,6 +59,38 @@ public class WSTrustInboundUtils {
         } catch (SecurityConfigException e) {
             // Error while adding WS Trust, we can't continue
             throw Utils.buildServerErrorResponse(e, "Error while adding WS-Trust configuration. " + e.getMessage());
+        }
+    }
+
+    public static WSTrustConfiguration getWSTrustConfiguration(InboundAuthenticationRequestConfig inboundAuth) {
+
+        String audience = inboundAuth.getInboundAuthKey();
+        try {
+            TrustedServiceData[] trustedServices =
+                    ApplicationManagementServiceHolder.getStsAdminService().getTrustedServices();
+
+            // TODO : check whether we need to throw an exception if we can't find a wstrust service
+            return Arrays.stream(trustedServices)
+                    .filter(trustedServiceData -> StringUtils.equals(trustedServiceData.getServiceAddress(), audience))
+                    .findAny()
+                    .map(trustedServiceData -> new WSTrustConfiguration()
+                            .audience(trustedServiceData.getServiceAddress())
+                            .certificateAlias(trustedServiceData.getCertAlias()))
+                    .orElse(null);
+
+        } catch (SecurityConfigException e) {
+            throw buildServerErrorResponse(e, "Error while retrieving wsTrust configuration for audience: " + audience);
+        }
+    }
+
+    public static void deleteWSTrustConfiguration(InboundAuthenticationRequestConfig inbound) {
+
+        try {
+            String trustedServiceAudience = inbound.getInboundAuthKey();
+            ApplicationManagementServiceHolder.getStsAdminService().removeTrustedService(trustedServiceAudience);
+        } catch (SecurityConfigException e) {
+            throw Utils.buildServerErrorResponse(e, "Error while trying to rollback wsTrust configuration. "
+                    + e.getMessage());
         }
     }
 }
