+++ components/org.wso2.carbon.identity.api.server.application.management/org.wso2.carbon.identity.api.server.application.management.v1/src/main/java/org/wso2/carbon/identity/api/server/application/management/v1/core/functions/application/inbound/custom/ApiModelToCustomInbound.java
@@ -34,7 +34,7 @@ public class ApiModelToCustomInbound implements
 
         InboundAuthenticationRequestConfig inboundRequestConfig = new InboundAuthenticationRequestConfig();
         inboundRequestConfig.setInboundAuthType(customInbound.getName());
-        inboundRequestConfig.setInboundAuthKey(customInbound.getInboundKey());
+        inboundRequestConfig.setInboundConfigType(customInbound.getConfigName());
         inboundRequestConfig.setProperties(getProperties(customInbound));
         return inboundRequestConfig;
     }
